requirements.txt
# 汽车销售 Agent 依赖包
# 用途：AI 安全课题实践，测试 LLM 攻击方法（Jailbreak、Prompt Injection 等）

# ==================== 核心依赖 ====================
# OpenAI Agents SDK
openai-agents>=0.11.0

# ==================== AI/ML ====================
pydantic>=2.0.0

# ==================== 向量检索 ====================
faiss-cpu>=1.7.0
numpy>=1.24.0

# ==================== 配置管理 ====================
# 内置模块，无需额外安装

# ==================== 工具 ====================
# 无额外依赖

# ==================== Python 版本 ====================
# 需要 Python 3.8 或更高
